DEX-677 Wrong order cancel
Tests fixes
