+++ dex/src/test/scala/com/wavesplatform/dex/model/OrderBookSpec.scala
@@ -13,7 +13,7 @@ import com.wavesplatform.dex.model.OrderBook.{LastTrade, Level, SideSnapshot, Sn
 import com.wavesplatform.dex.settings.MatchingRule
 import com.wavesplatform.dex.settings.OrderFeeSettings.{DynamicSettings, OrderFeeSettings}
 import com.wavesplatform.dex.time.NTPTime
-import com.wavesplatform.dex.{Matcher, MatcherSpecBase, NoShrink}
+import com.wavesplatform.dex.{MatcherSpecBase, NoShrink}
 import org.scalacheck.Gen
 import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.should.Matchers
@@ -458,7 +458,7 @@ class OrderBookSpec
 
       val maker = limit(mAmt, SELL, orderMFee)
       val taker = if (isTMarket) market(tAmt, BUY, normalizedOrderTFee, tFeeAsset) else limit(tAmt, BUY, normalizedOrderTFee, tFeeAsset)
-      val gmtf  = Matcher.getMakerTakerFee(ofs)(_, _)
+      val gmtf  = Fee.getMakerTakerFee(ofs)(_, _)
       val evt   = { ob.add(maker, System.currentTimeMillis, gmtf); ob.add(taker, System.currentTimeMillis, gmtf) }.head
 
       evt shouldBe a[OrderExecuted]
